Move RoleRuleProvider to RoleCondition
